Introduce composable useInputReconciler hook
