+++ packages/slate-react/src/hooks/input-reconciler/clipboard/index.ts
@@ -1,0 +1,1 @@
+export { useClipboardReconciler } from './use-clipboard-reconciler'
