+++ packages/slate-react/src/hooks/input-reconciler/composition/before-input/index.ts
@@ -1,0 +1,1 @@
+export { useBeforeInputReconciler } from './use-before-input-reconciler'
