+++ packages/slate-react/src/hooks/input-reconciler/composition/index.ts
@@ -1,0 +1,1 @@
+export { useCompositionReconciler } from './use-composition-reconciler'
