+++ packages/slate-react/src/hooks/input-reconciler/composition/use-composition-reconciler.ts
@@ -1,0 +1,4 @@
+import { useBeforeInputReconciler } from './before-input'
+
+// To-do: Conditional composition reconciler based on browser capabilities
+export const useCompositionReconciler = useBeforeInputReconciler
