+++ packages/slate-react/src/hooks/input-reconciler/drag-and-drop/index.ts
@@ -1,0 +1,1 @@
+export { useDragAndDropReconciler } from './useDragAndDropReconciler'
