+++ packages/slate-react/src/hooks/input-reconciler/focus/index.ts
@@ -1,0 +1,1 @@
+export { useFocusReconciler } from './use-focus-reconciler'
