+++ packages/slate-react/src/hooks/input-reconciler/index.ts
@@ -1,0 +1,1 @@
+export { useInputReconciler } from './use-input-reconciler'
