+++ packages/slate-react/src/hooks/input-reconciler/selection/index.ts
@@ -1,0 +1,1 @@
+export { useSelectionReconciler } from './use-selection-reconciler'
